`KafkaConsumerConfig` add `SwiftKafka` options
Motivation:

We want to use the `KafkaConsumerConfig` to determine if the
`KafkaConsumer` is part of a consumer-group (subscription) or assinged
to a single partition of a topic (assingment). This should also be
implemented in a way that errors are caught at compile time.
Furthermore, we want users to be able to determine a
`BackPressureStrategy` for their respective `KafkaClient`.

Modifications:

* add two new configuration options to `KafkaConsumerConfig`:
  `ConsumptionStrategy` and `BackPressureStrategy`
* remove `KafkaConsumerConfig`'s `groupID` option, this is now taken
  care of by the `consumptionStrategy` option, which is better as this
  ensures at compile time that now groupID mismatch errors occur
* rename `KafkaProducerConfig` back to `ConfigEnums` as it is not really
  meant for shared configurations, rather as a namespace for _all_
  configuration options
* update README