Create `KafkaProducer` with factory method
Motivation:

By creating the `KafkaProducer` with a factory method that also returns
the `KafkaAsyncSequence` for receiving message acknowledgements, we
ensure that `KafkaAsyncSequence` will eventually get terminated as the
user now owns the memory reference to it.
It is important that the `KafkaAsyncSequence` gets terminated because
otherwise the `KafkaProducer`'s run task runs forever.

Modifications:

* make initializer of `KafkaProducer` `private`
* create new factory mehtod `KafkaProducer.newProducer` for initializing
  the `KafkaProducer` alongside the acknowledgement `KafkaAsyncSequence`
* update tests
* update README