KafkaProducer.shutdownGracefully kill poll loop
Motivation:

Previously, we had to wait for the `KafkaProducer.acknowledgements`
`AsyncSequence` delegate method `didTerminate()` to be invoked in order
to shutdown the poll loop. This was an issue as this lead to the poll
loop never stopping when e.g. the user never subscribed to the
`acknowledgements` `AsyncSequence` and therefore the `didTerminate()`
delegate method would never be invoked.

Modifications:

* make `KafkaPollingSystem.terminate()` `internal`
* invoke `pollingSystem.terminate()` in
  `KafkaProducer.shutdownGracefully`
* remove poll loop shutdown workaround in test

Result:

After your change, what will change.