Review Franz
Motivation:

Instead of storing a messages offset after every yield to the
`messages` `AsyncSequence`, we want to do that after every invocation to
the `message` `AsyncSequenceIterator`'s `next()` method to ensure that
the user has actually received the message. Yielding only enqueued a
message but did not ensure that the user actually consumed it through
the `AsyncSequence`.

Modifications:

* `KafkaConsumerConfig`
    * revert to plain `librdkafka` `enable.auto.commit` implementation
* `KafkaConsumer`
    * use `rd_kafka_offsets_store()` instead of `rd_kafka_commit` to store a
      message's offset after it has been consumed through the `messages`
      `AsyncSequence` -> more lightweight than `rd_kafka_commit` + we
      can still leverage `librdkafka`'s auto-commit logic
    * commit message offsets on `ConsumerMessagesAsyncIterator.next()`
      instead of after yielding (explaination in Motivation section)