`KafkaProducer`: events- instead of acks sequence (#96)
* `KafkaProducer`: event instead of acks sequence

Motivation:

`KafkaProducer`: we want to expose a general `KafkaProducerEvent` type
in the `AsyncSequence` that was formerly just for message
acknowledgements. This enables us to add more events such as statistics
in the future.

The reason behind having a single `AsyncSequence` for all event types
instead of having separate `AsyncSequence`s for each event type is that
we need to ensure that all events that we subscribe to in `librdkafka`
are actually consumed by the `AsyncSequence` in our `KafkaProducer`
type. Otherwise we could run out of memory.

Now by giving the user the entire events `AsyncSequence`, they decide if
they want to consume event or drop it.

> **Note**: Logs will be consumed regardless of the
> `KafkaProducerEvents` `AsyncSequence`

Modifications:

* create a new `enum` `KafkaProducerEvent`
* rename `KafkaMessageAcknowledgements` -> `KafkaProducerEvents`
* rename `KafkaProducer.makeProducerWithAcknowledgements` ->
  `.makeProducerWithEvents`
* update tests
* update README

* Review Franz

Modifications:

* fix documentation typos
* `KafkaProducerEvent`: replace factory method with `init`
* create new type `KafkaProducerMessageStatus` effectively representing
  if a `KafkaProducerMessage` was acknowledged or not
* fix README nit

* `KafkaDeliveryReport` type

Modifications:

* create new type `KafkaDeliveryReport` containing a message's status
  and its id
* remove ID from `KafkaAcknowledgedMessage`
* remove `KafkaAcknowledgedMessageError`
* rename `KafkaProducerEvent.deliverReport` -> `.deliverReports`
* update README
* update tests

* Fix documentation nits

* Review Gus

Modifications:

* fix typos
* rename `KafkaProducerEvents.KafkaProducerEventAsyncIterator` ->
  `KafkaProducerEvents.AsyncIterator`

* Fix typo

* Fix (letter) case nit

Co-authored-by: Franz Busch <privat@franz-busch.de>

---------

Co-authored-by: Franz Busch <privat@franz-busch.de>