Commit messages by topic, partition, offset
Motivation:

In some cases, for instances when messages are handed out
to other processes, it is not reasonable to hold on to the
whole Kafka message to be able to commit it at a later time.
It consumes much less memory to hold only onto the details
needed to commit the message.

Modifications:

Add new public methods that allow users to commit messages
if they have a reference to the topic, partition, and offset
instead of only the whole message.

Result:

Developers no longer need to keep the whole message around to
commit it, only the topic, partition, and offset.