Run consume loop 1000 times each
Changes the benchmark to iterate over the consume loop a total of 1000
times. We now produce 1000*1000 (1e6) messages into the topic, and on
every benchmark iteration consume 1000 messages each.

I had to comment the `librdkafka_with_offset_commit_messages_*`
benchmark. For some reason, the benchmark suite keeps re-running it, and
I have seen occasional failures when attempting to commit offsets. It's
unclear to me why that happens right now, but decided it's not worth the
investigation at the moment.